% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a0_data_no_zeros}
\alias{a0_data_no_zeros}
\title{The initial amplitude data, for both time series, with the zeros removed.}
\format{a 1473 x 4 data.frame}
\usage{
a0_data_no_zeros
}
\description{
Note that the removal of the zero time points (due to filtering) means the
indices don't line up with the Crux* objects, if you want to get certain time
points, match by date (coerce both via as as.POSIXct). If you really wanted
to you could match via rownames (they technically still match).
}
\keyword{datasets}
