% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-amplitude.R
\name{a0_sample_generator}
\alias{a0_sample_generator}
\title{Sample the PPD for the initial amplitude model}
\usage{
a0_sample_generator(sea_surface_height, n_samples = 500)
}
\arguments{
\item{sea_surface_height}{the sea surface height}

\item{n_samples}{defaults to 500, how many samples to return}
}
\value{
vector of legnth n_samples
}
\description{
Currently only works in settings where 
\code{\link[parallel]{mclapply}} functions.
}
