% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{approx_post_chol}
\alias{approx_post_chol}
\title{The cholesky needed to approximately sample the posterior around a 
curve.}
\format{a  6 x 6 lower triangular matrix}
\usage{
approx_post_chol
}
\description{
This cholesky is found by getting the correlation matrix for the mu_{beta}
samples, and combining it with the average (over time) standard deviation
for each of the coefficients
}
\keyword{datasets}
