% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Crux_KP150_Phs1}
\alias{Crux_KP150_Phs1}
\title{First 6 months of data}
\format{A list including 11 pressures, and ~737 measurements there of}
\usage{
Crux_KP150_Phs1
}
\description{
Subtitle
}
\keyword{datasets}
