% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dated_filename}
\alias{dated_filename}
\title{Date a filename}
\usage{
dated_filename(filename, folder_path = "./plots/", date = Sys.Date(),
  sep = "_")
}
\arguments{
\item{filename}{the filename, with file extensions}

\item{folder_path}{the directory for the file, defaults to "./plots/"}

\item{date}{a date string, defaults to Sys.Date()}

\item{sep}{seperator character, if set to "" and combined with a leading "/"
in `filename`, then dated directories will be created.}
}
\value{
a filepath/filename string
}
\description{
This function will create the directories in the full file path it 
constructs, and return the full file path to the specified file.
}
