% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddcurves2.R
\docType{package}
\name{ddcurves2-package}
\alias{ddcurves2-package}
\alias{rstanlm}
\title{Depth density curves using R/Stan}
\description{
The \pkg{ddcurves2} package is a collection of models for oceanographic 
depth-density data
}
