% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double-tanh-plotter.R
\name{double_fitted_curve_plotter}
\alias{double_fitted_curve_plotter}
\title{Plot the fitted curves for the double tanh model.}
\usage{
double_fitted_curve_plotter(stan_obj, index, stan_dat, z_new = seq(from =
  -300, to = 0, length.out = 301), plot_pred_intervals = FALSE, ...)
}
\arguments{
\item{stan_obj}{a 'stanfit' object for a model with two tanh components}

\item{index}{a time point in the original data set}

\item{stan_dat}{the data set}

\item{z_new}{a sorted vector of z (x) values to plot the curve over}

\item{plot_pred_intervals}{if TRUE, also plot the prediction intervals in
addition to the mean credible intervals}

\item{...}{currently unused}
}
\value{
a ggplot2 object of the fitted curve, split by chain, with credible
interval
}
\description{
Plots the fitted curves for a single time index, for each chain.
THIS FUNCTION IS VERY FRAGILE, AND IS LIKELY TO BREAK IF YOU CHANGE ANYTHING
IN THE MODEL STRUCTURE (BECAUSE IT REORDERS THE PARAMETERS).
}
\details{
You will know if it breaks because the plot will look really whacky.
}
