% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double-tanh-no-time-series.R
\name{double_tanh_no_timeseries}
\alias{double_tanh_no_timeseries}
\title{Fit the double tanh model, with no timeseries structre and an simple
hierarchical mean}
\usage{
double_tanh_no_timeseries(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{Stan data list, with elements:
n_depths (the number of depth values which we record density at)
n_times (the number of time points at which density is recorded)
depths (vector of depth values)
densities (an n_times * n_depths matrix of density observations.)}

\item{...}{other parameters that get passed to 
\code{\link[rstan:stanmodel-method-sampling]{sampling}}}
}
\value{
an object of class 'stanfit'
}
\description{
Fit the double tanh model, with no timeseries structre and an simple
hierarchical mean
}
