% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double-tanh-time-series.R
\name{double_tanh_with_timeseries}
\alias{double_tanh_with_timeseries}
\title{Fit the double tanh model, with time series structure}
\usage{
double_tanh_with_timeseries(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{Stan data list}

\item{...}{other parameters that get passed to \link[rstan]{sampling}}
}
\value{
an object of class 'stanfit'
}
\description{
Fit the double tanh model, with time series structure
}
