% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indep_fit_chol_emp_cor_mat}
\alias{indep_fit_chol_emp_cor_mat}
\title{Cholesky decomposition of the empirical correlation matrix of the fully 
independent, single tanh fits.}
\format{an 11 x 11 lower triangular matrix, suitable for passing directly
into Stan's "cholesky_factor_corr" type.}
\usage{
indep_fit_chol_emp_cor_mat
}
\description{
For the First 6 months of data, "Crux_KP150_Phs1"
}
\keyword{datasets}
