% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{init_jitter}
\alias{init_jitter}
\title{Init jitterer}
\usage{
init_jitter(init_list, jitter_magnitude = 0.001)
}
\arguments{
\item{init_list}{A list of initial values.}

\item{jitter_magnitude}{Magnitude of the jittering.}
}
\value{
A jittered version of \code{init_list}.
}
\description{
Jitters the initial values for a list of initial values that is in the form
required by \code{\link[rstan:stanmodel-method-sampling]{sampling}}. 
\code{jitter_magnitude} controls the magnitude of the jittering, i.e. An 
initial value of 1 with \code{jitter_magnitude} = 0.001 will get jittered to
be uniformly distributed on the interval between 1 +/- (1 * 0.001 / 2).
}
