% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-beta-diag.R
\name{mean_beta_density_diag}
\alias{mean_beta_density_diag}
\title{Generate density plots by chain of the mean's of the betas.}
\usage{
mean_beta_density_diag(stan_obj, print = FALSE, print_dir = "", ...)
}
\arguments{
\item{stan_obj}{stan model fit}

\item{print}{logical, if TRUE and print_dir specified, print a pdf to print_dir}

\item{print_dir}{directory to put image in}

\item{...}{optional params passed to pdf()}
}
\value{
ggplot2 object from bayesplot
}
\description{
Generate density plots of the Mean's of the Betas by chain
}
