% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma-diag.R
\name{mean_sigma_trace_diag}
\alias{mean_sigma_trace_diag}
\title{Generate traceplots of all parameters with 'sigma' in their name.}
\usage{
mean_sigma_trace_diag(stan_obj, print = FALSE, print_dir = "", ...)
}
\arguments{
\item{stan_obj}{stan model fit}

\item{print}{logical, if TRUE and print_dir specified, print a pdf to print_dir}

\item{print_dir}{directory to put image in}

\item{...}{optional params passed to pdf()}
}
\value{
ggplot2 object from bayesplot
}
\description{
Generate traceplots of all parameters with 'sigma' in their name.
}
