% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_model.R
\name{naive_model}
\alias{naive_model}
\title{Fit the naive (single mean, no time dependency) stan model}
\usage{
naive_model(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{Stan data list}

\item{...}{other parameters that get passed to \link[rstan]{sampling}}
}
\value{
an object of class 'stanfit'
}
\description{
Fit the naive (single mean, no time dependency) stan model
}
