% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-predict-new.R
\name{plot_post_new_points}
\alias{plot_post_new_points}
\title{produce a plot of the approximate posterior}
\usage{
plot_post_new_points(new_depths, new_densities, plot_quantile = 0.025,
  ...)
}
\arguments{
\item{new_depths}{new_depths}

\item{new_densities}{new_depths}

\item{...}{arguments passed to  \link{sample_new_post}}
}
\value{
a ggplot object
}
\description{
produce a plot of the approximate posterior
}
