% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-plotters.R
\name{plot_spline_fit}
\alias{plot_spline_fit}
\title{Plot the fittedn curves from the spline model}
\usage{
plot_spline_fit(model_fit, stan_data_list, times_to_plot = 1:4,
  split_by_chain = TRUE, interval_quantile = 0.025, add_sigma_y = FALSE)
}
\arguments{
\item{model_fit}{a "stanfit" object of the latent spline model}

\item{stan_data_list}{the stan data list associated with the model_fit}

\item{times_to_plot}{a vector of time points to plot the fits for.}

\item{split_by_chain}{should the plots be split by chain? If TRUE, then
they will be, if FALSE the chains will be combined}

\item{interval_quantile}{what is the lower quantile of the interval of 
interest (0.025 = 95% intervals)}
}
\value{
a ggplot2 object
}
\description{
Plot the fittedn curves from the spline model
}
