% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-plotters.R
\name{plot_spline_posteriors}
\alias{plot_spline_posteriors}
\title{Plot Spline Posteriors}
\usage{
plot_spline_posteriors(model_fit, stan_data_list, split_by_chain = TRUE,
  interval_quantile = 0.025)
}
\arguments{
\item{model_fit}{a "stanfit" object of the latent spline model}

\item{stan_data_list}{the stan data list associated with the model_fit}

\item{split_by_chain}{should the plots be split by chain? If TRUE, then
they will be, if FALSE the chains will be combined}

\item{interval_quantile}{what is the lower quantile of the interval of 
interest (0.025 = 95% intervals)}
}
\value{
a ggplot2 object? maybe something that has been grid.arranged?
 I think I can do this with faceting.
}
\description{
Plot the spline posteriors over time, potentially split by chain
}
