% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-residual.R
\name{resid_depth_plot}
\alias{resid_depth_plot}
\title{Plot the Residual values (y) as a function of Depth (x)}
\usage{
resid_depth_plot(stan_data_list = stan_dat, stan_obj = model_fit, ...)
}
\arguments{
\item{stan_data_list}{the data list passed to the stanmodel, for structure
of this list, see \code{\link{double_tanh_no_timeseries}}}

\item{stan_obj}{the stanfit object fitted using the aforementioned data}

\item{...}{nused, possible graphical options}
}
\value{
ggplot object of the specified residual plot
}
\description{
Plot the Residual values (y) as a function of Depth (x)
}
