% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted-residual.R
\name{resid_fitted_plot}
\alias{resid_fitted_plot}
\title{Plot the residual values (y) against the fitted values (x)}
\usage{
resid_fitted_plot(stan_data_list = stan_dat, stan_obj = model_fit, ...)
}
\arguments{
\item{stan_data_list}{the data list passed to the stanmodel}

\item{stan_obj}{the stanfit object fitted using the aforementioned data}

\item{...}{unused, possible graphical options}
}
\value{
ggplot object of the residual plot
}
\description{
Plot the residual values (y) against the fitted values (x)
}
