% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-predict-new.R
\name{sample_new_post}
\alias{sample_new_post}
\title{Approximately sample the posterior for a new set of depths and densities}
\usage{
sample_new_post(new_depths, new_densities, n_samples = 1000,
  nlm_inits = c(1020, 2, 75, 45, 150, 47))
}
\arguments{
\item{new_depths}{new depths at which new_densities is measured}

\item{new_densities}{new density measurements}

\item{n_samples}{number of approximate samples to draw. In some specific
circumstances this can be different to the number of samples returned}

\item{nlm_inits}{initial values for nlm (the minimiser), defaults are 
reasonable for data on the original scale.}
}
\value{
list with two components, the first being the mean estimate, and the
second the samples around this mean
}
\description{
Approximately sample the posterior for a new set of depths and densities
}
