% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tanh_fixed_corr_structure_normal_mean.R
\name{singe_tanh_fixed_corr_weak_mean}
\alias{singe_tanh_fixed_corr_weak_mean}
\title{Sample the single tanh model, with fixed correlation structure and weak (les
tight bounds) normal mean on the beta_zero parameter}
\usage{
singe_tanh_fixed_corr_weak_mean(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{list of data to pass into stan, including cholesky
of the empirical correlation structure}

\item{...}{additional arguments to pass to \link[rstan]{sampling}}
}
\value{
stanfit object
}
\description{
Sample the single tanh model, with fixed correlation structure and weak (les
tight bounds) normal mean on the beta_zero parameter
}
