% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tanh_fixed_corr_mat.R
\name{single_tanh_fixed_corr_mat}
\alias{single_tanh_fixed_corr_mat}
\title{Fit the Single hyperbolic tangent and with fixed correlation matrix.}
\usage{
single_tanh_fixed_corr_mat(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{Stan data list}

\item{...}{other parameters that get passed to \link[rstan]{sampling}}
}
\value{
an object of class 'stanfit'
}
\description{
Fit the Single hyperbolic tangent and with fixed correlation matrix.
}
