% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tanh_full_covariance.R
\name{single_tanh_full_covariance}
\alias{single_tanh_full_covariance}
\title{Fit the Single hyperbolic tangent and full covariance structure model.}
\usage{
single_tanh_full_covariance(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{Stan data list}

\item{...}{other parameters that get passed to \link[rstan]{sampling}}
}
\value{
an object of class 'stanfit'
}
\description{
Fit the Single hyperbolic tangent and full covariance structure model.
}
