% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_tanh_no_timeseries.R
\name{double_tanh_no_timeseries}
\alias{double_tanh_no_timeseries}
\title{Fit the double tanh model, with no timeseries structre and an simple
hierarchical mean}
\usage{
double_tanh_no_timeseries(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{Stan data list}

\item{...}{other parameters that get passed to \link[rstan]{sampling}}
}
\value{
an object of class 'stanfit'
}
\description{
Fit the double tanh model, with no timeseries structre and an simple
hierarchical mean
}
