% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-amplitude.R
\name{initial_amplitude_model}
\alias{initial_amplitude_model}
\title{Fit the initial amplitude model}
\usage{
initial_amplitude_model(stan_data_list, ...)
}
\arguments{
\item{stan_data_list}{Stan data list}

\item{...}{other parameters that get passed to \link[rstan]{sampling}}
}
\value{
an object of class 'stanfit'
}
\description{
Fit the initial amplitude model
}
